Implement api/users route
